Nest ScheduleItems within Rooms
Fixes a lot of specs. It turned out that objects persisted to the test
database by assigning them to instance variables instead of using let
and let!, never get removed from the database. There also is a bug where
after FactoryGirl creates an objects, its associations do not get
reloaded automatically, so that had to be added to the factories.