Refactor views
Up to this point I was not aware that there is a way to insert new lines
between arguments in a function call (it's only possible when a line
ends with a comma). For some inexplicable reason I thought that empty
lines between tags would produce a br tag, I'm glad I found out I was
wrong.