Change the schedule items dimensions algorithm
Previously, dimensions of a schedule item  were calculated based
solely on the number of other schedule items that were going on during
the whole time the item is going on, with disregard to the fact that
those other items overlap with each other or not. This resulted in
item's width being narrower than it could be. Now the algorithm bases
width and left position of the maximum number of items going on at the
same time. Left position is given prority to items that have a longer
remaning time firstly, and room name secondly.